[#5384] Cleanup code related to cluster
Create a StreamChannel interface to enhance scalability.
